requirements.txt
streamlit
streamlit_autorefresh
fpdf
python-dotenv
pandas
openpyxl
os
json
random
shutil
datetime
ZipFile